More refactoring to support new file structure
