+++ src/Admin/Admin.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Admin = () => <>Admin</>;
+
+export default Admin;
