+++ src/App.tsx
@@ -2,10 +2,11 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
 import { useApiGet, TApiResponse } from "./hooks/useApiHook";
-import Home from "./Restaurants/Landing/Home";
+import Home from "./Restaurants/Landing/Landing";
 import Restaurants from "./Restaurants/Restaurants";
 import Menu from "./Restaurants/Menu/Menu";
 import Order from "./Restaurants/Order/Order";
+import Admin from "./Admin/Admin";
 
 function App() {
   // call to the hook
@@ -24,7 +25,7 @@ function App() {
           <Route path="/menu" element={<Menu />} />
           <Route path="/order" element={<Order />} />
         </Route>
-        <Route path="/admin" element="ADMIN" />
+        <Route path="/admin" element={<Admin />} />
       </Routes>
     </Router>
   );
